csm/csm_pc_tests/requirements.txt
async-timeout==4.0.3
bcrypt==4.0.1
cachetools==5.3.1
cffi==1.15.1
cryptography==41.0.4
decorator==5.1.1
Deprecated==1.2.14
et-xmlfile==1.1.0
fabric==3.2.2
future==0.18.3
ifaddr==0.2.0
invoke==2.2.0
openpyxl==3.0.9
paramiko==3.3.1
pathlib2==2.3.7.post1
pefile==2023.2.7
py2exe==0.13.0.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
scp==0.14.5
six==1.16.0
tenma-serial==1.1.1
wrapt==1.15.0
zeroconf==0.112.0
